requirements.txt
Django
djangorestframework
python-decouple
django-filter
djangorestframework-simplejwt
drf-spectacular
requests

